Increase default for `max_tokens`.
This increases the default to 2048, while also increasing the max.
LLaMa2's context window is 4096, which is why I chose that #. That
said I'm not sure if that makes sense or not. In theory:

- the context window should dictate the size of the inputs/prompt;
  anything that overflows 4096 tokens presumably needs to be
  compressed/summarized

- the output maximum is a function of how much GPU memory is available
  after prefilling the input tokens; which means the maximum value is
  again a function of the input length.

We'll do something smarter later, for now this should prevent the model
from being "cut off" as often.