Paginate listing messages (#20)
This paginates the `/messages` API response.

To do this I added a `/v3` endpoint, since this is a non-backwards
compatible change. I was pleased how this came together.

After this lands I'll update the UI to use the `/v3` endpoints. Then I
can follow up by removing the `v2` code.