Support for login URLs (#23)
This adds support for login URLs, which are intended to give non AI2
folks temporary access to the demo.

To generate a login URL, an administrator will submit a POST request to
the API:

```
curl -X POST -d '{ "expires_in": "7d", "client": "sam.skjonsberg@gmail.com" }' ...
```

The administrator will be authenticated via our existing mechanism,
which uses an HTTP cookie and/or bearer token. The administrator's email
must be in a list of configured administrators, which will be limited to
ReOps.

Once the login URL is generated, it can be clicked once. The client who
clicks the URL will be given a authentication token with a 7 day expiry.
They must click this before the login URL expires (which is controlled
via the `expires_in` value given at creation time).

To accomplish this I extended the `client_token` table. All records now
have a token type. Tokens of `"login"` type can only be used to login
via URL. Tokens of `"client"` type can only be used to login via a HTTP
cookie/bearer token.

I also added a `NULL`able column which tracks who created a token. This
will only be set for `"login"` tokens.