Michalg/login endpoint (#31)
Currently, requests for https://olmo-api.allen.ai/v3/whoami redirect to
Skiff Login. Once authn/authz is completed, requests for
https://olmo-api.allen.ai/v3/whoami are satisfied with this code:

```
    def whoami(self):
        agent = self.request_agent()
        if agent is None or agent.expired():
            # Use NGINX mediated auth; see https://skiff.allenai.org/login.html
            email = request.headers.get("X-Auth-Request-Email")
            if email is None:
                raise exceptions.Unauthorized()

            agent = self.dbc.token.create(email, token.TokenType.Auth, timedelta(days=7))

        return self.set_auth_cookie(jsonify(AuthenticatedClient(agent.client)), agent)
```

In this code, we make a new token based on your Skiff Login identity if
you don't have one, or return the current one you have.

Unfortunately, the Olmo UI at https://olmo.allen.ai/ makes Javascript
requests for https://olmo-api.allen.ai/v3/whoami to decide if the user
is authorized or not. If authorized, the UI renders the content (by
making additional requests to olmo-api). If unauthorized, the UI renders
"unauthorized". But the catch is that since requests for
https://olmo-api.allen.ai/v3/whoami can redirect to Skiff Login (in case
of a new browser session), instead of just returning authorized or
unauthorized, the Javascript requestor gets confused because it can't
execute a full Google login workflow.

To make the Olmo UI work, we need to do this:

- Update Olmo API so requests for https://olmo-api.allen.ai/v3/whoami do
not require Skiff Login, and either return authorized or unauthorized,
without the side-effect of redirecting to Skiff Login.
   - That's this change
- Update Olmo UI so that an unauthorized request for
https://olmo-api.allen.ai/v3/whoami sends the browser into an
authorization flow and returns to the Olmo UI homepage.
   - The new authorization flow is now /v3/login (in this change)
   - The update to the UI will be a different change

---------

Co-authored-by: Sam Skjonsberg <sams@allenai.org>