Undelete prompt templates
The focus of this PR is on adding the ability to undelete/unarchive
a prompt template.

To do so I added the `deleted` boolean to `PATCH /templates/prompt/<id>`.

I kept `DELETE /templates/prompt/<id>` even though it's unnecessary now.
This is both in the interest of API backwards compatibility and
consistency.

As part of this I also deprecated the `author` field and started using
`creator` instead. I kept the DB schema unchanged, and am setting both
fields as to not break the client.

https://github.com/allenai/olmo-api/issues/46