Add API support for authenticating after logging in.
This adds the `redirect` query string argument to the `/v3/login/skiff`
endpoint.

By setting this the client can direct redirection after successful
authentication.

If the redirect URL isn't to the canonical UI origin (per the server's
configuration) a HTTP 400 Bad Request is returned.

This is part 1 of https://github.com/allenai/olmo-ui/issues/82. After
this change has been released I can modify the UI to set the
`?redirect=...` argument when prompting the user to login.