Add ability to sort labels
This adds the ability to change the sort direction of the returned
labels.

I tried to lay some groundwork for adding sort capabilities to other
endpoints, but chose not to implement those paths for now. As part of
this I revised the error handling so that those errors get mapped to
explicit HTTP codes and messages instead of printing stack traces.
This doesn't mean the exception messages are safe, but we have
a ticket tracking efforts to review those exceptions that we'll take
care of before going public.