Incognito messages (#74)
This adds support for setting an `incognito` property when creating
messages. Messages that have this value set to `true` are only visible
to that user (and administrators). This prevents these messages from
showing up globally to all users, which is helpful for
red-teaming/evaluating toxicity.

As part of this I made the completions endpoint admin-only. It's not
currently used by the UI, and trying to implement and verify the
visibility constraints I'm adding in this PR didn't feel worth it.
Particularly since this endpoint is one that should be going away soon,
once model I/O storage is something InferD does.

Oh, and as part of this I fixed how local administrators work.
Previously we used the settings from prod, inherited via
`./bin/bootstrap`. Now we use a new, local value to distinguish the two
environments more clearly.

I'll post some details in a bit about what the API looks like, as it's a
lot of text due to the nature of how the API works.

https://github.com/allenai/olmo-api/issues/70