Add logprobs (#76)
This plumbs through logprobs, which were added to the InferD Tulu2
worker in
https://github.com/allenai/inferd-tulu2/commit/f50ac6a256bcde1c8d68d5f366f92e8680f11fd7.

One side-effect of this PR was to change usages of `Json()` to
`Jsonb()`. The prior appeared to work in all cases except where a
`COALESCE` is used, probably because it causes `psycopg` to add a cast.
`Jsonb()` is more correct, as that's what I decided to use at the schema
level. So this is a good thing, I think.

This should wrap up https://github.com/allenai/olmo-api/issues/55.