Better handling of rpc errors (#81)
This modifies the API to handle cases where the InferD stream fails b/c
of a generic RPC error. Right now an exception occurs but is not
propagated to the UI. This modifies the API to handle the exception,
inform the UI and fail more gracefully.

Currently this occurs when a prompt exceeds the limit. Right now InferD
doesn't [terminate the
stream](https://github.com/allenai/inferd/issues/508), which means
something else (eventually) does. It's probably something `redis` is
doing (or maybe some sort of default gRPC deadline). When that occurs it
manifests as an `RpcError` that indicates that the [`stream has been
removed`](https://github.com/allenai/olmo-ui/issues/3#issuecomment-1858575731).

This code changes the API to handle that case gracefully. Instead of
abruptly terminating things we'll send an error to the UI and
"finalized" the message so that things aren't left dangling.