Use `finish_reason` to detect and handle errors
This modifies the API to support the newly added `finish_reason`
and use it to detect scenarios where things failed mid-generation.

Importantly this allows us to show an error when the prompt is too
long. This resolves https://github.com/allenai/olmo-ui/issues/3 after
several weeks worth of work.

The error that's displayed in the UI isn't great, I cut
https://github.com/allenai/olmo-ui/issues/151 to improve that. But
displaying an error is better than an indefinitely loading spinner.

As part of this I also modified a few defaults so that we don't
send `logprobs` or `stop` unless they're specified by the client
and modified `OutpartPart` to capture that `logprobs` might be `None`.