No-HTML datachips
This modifies datachips so that we don't use HTML markup for injecting
them into content. We can't do so because people might enter their
own HTML, and the model  might emit HTML -- which means we need to use
a markup mechanism that's *less* likely to collide w/ the user/model I/O.

This PR modifies the markup mechanism that that datachips can be
referenced like so:

  `:creator/name`

The `:creator/name` is collectively called a `ref`, which we add as a
new, globally `UNIQUE` column.

To keep things simple:

- I plan to `DROP TABLE datachip` before applying the migrations;
  there's only a few datachips now, none of which are important to
  preserve

- The `UNIQUE` constraint applies to deleted datachips as well. I
  didn't want to have to reason about allowing name re-use and the
  edge-cases that are introduced.

https://github.com/allenai/olmo-api/issues/64