No-HTML datachips (#85)
This modifies datachips so that we don't use HTML markup for embedding
them in messages. We can't do so because people might enter their own
HTML, and the model might emit HTML -- which means we need to use a
markup mechanism that's *less* likely to collide w/ the user/model I/O.

This PR modifies the markup mechanism so that datachips can be
referenced like so:

  `:creator/name`

The `:creator/name` is collectively called a `ref` whose type if
`DatasetRef` (which is just an alias for `str`). The `ref` field is
added as a new, globally `UNIQUE` and `NOT NULL` column. I elected to
use this instead of IDs to make them friendlier/more human readable. If
we were to use IDs things would look like:

```
Please summarize this :dc_12kdb3hg2
```

Instead, they'll look like:

```
Please summarize this :murphy@allenai.org/paper
```

To keep things simple:

- I plan to `TRUNCATE datachip` before applying the migrations; there's
only a few datachips now, none of which are important to preserve

- The `UNIQUE` constraint applies to deleted datachips as well. I didn't
want to have to reason about allowing name re-use and the edge-cases
that are introduced.

https://github.com/allenai/olmo-api/issues/64