Add model type to model config to support base model (#91)
This adds a new `model_type` field for each model that can either be
`base` or `chat`. This is to support the base model in the OLMo UI
(@arnavic is there a ticket for this?). See
https://github.com/allenai/olmo-ui/pull/195 for prior UI work that
supports parsing this field.

@arnavic storing and surfacing the model type (for old messages) is a
bit more work and probably warrants a bit of discussion on the API side.
Is this something we can hold off on?

I plan to update the config secret to us the following settings for
InferD:
```json
"inferd": {
  "address": "grpc-inferd.allen.ai:443",
  "token": "<TOKEN>",
  "default_model": "olmo-7b-chat",
  "available_models": [
    {
      "id": "olmo-7b-chat",
      "name": "OLMo 7B - Chat",
      "description": "AI2's 7B model trained on the Dolma dataset and fine-tuned for chat.",
      "compute_source_id": "csc_01hnf2b26vq2p0ag97aw6bznjr",
      "model_type": "chat"
    },
    {
      "id": "olmo-7b-base",
      "name": "OLMo 7B - Base",
      "description": "AI2's 7B model trained on the Dolma dataset.",
      "compute_source_id": "csc_01hpmr468ksmbbfab3szhpjyws",
      "model_type": "base"
    },
    {
      "id": "tulu2",
      "name": "Tulu2",
      "description": "A 70B parameter model that is a fine-tuned version of Llama 2.",
      "compute_source_id": "csc_01hj45x4704wjy6en5vvx7ex7c",
      "model_type": "chat"
    }
  ]
},
```