Pull message blueprint out of v3 blueprint (#99)
Made this PR as part of the leadup to adding togetherAI support. Wanted
to make the v3 blueprint smaller before changing things.

This is a fairly 1:1 removal of anything that goes under the `/message`
path. The main changes that were made were renaming things and pulling
more complicated code out to a function.

All e2e tests pass