Abstract out the inference engine so we can use different ones (#100)
Closes #97 

This PR adds an `InferenceEngine` `Protocol` that allows us to
substitute different implementations of an inference engine. It includes
an InferD implementation of that protocol as well.

I used a Python formatter called Ruff while working on this, so there
will be a decent amount of formatting changes. Apologies for that! I'll
try to highlight important changes in files where that happened.

All E2E tests pass.