tinkering with request contexts (#104)
I was curious about using
[contexts](https://flask.palletsprojects.com/en/3.0.x/appcontext/) to
help remove some of the parameter passing we need to do for our
application right now. I've moved the config (and available_models) to a
`LocalProxy` based on the advice from the flask docs:
https://flask.palletsprojects.com/en/3.0.x/appcontext/#storing-data

It seems to work well and removed the config passing for our blueprints!
What do you all think about it?