Auth0 Foundation (#108)
This is a very basic POC of implementing auth0 and getting the token
from the UI.

It uses the Authlib library and Auth0's implementation from their
quickstart. Most code for this can be found in `auth0.py`. There is some
code that extracts the user's token in the `whoami` handler in
`src/user/__init__.py` but doesn't do anything with it.

Corresponds with https://github.com/allenai/olmo-ui/pull/414