Change regular expression to avoid matching URL (#115)
## Description

Closes: https://github.com/allenai/olmo-ui/issues/457

## Review Instruction
Base on my debugging it looks like MG text contains a url with github
link and in parse.py we have a code to look for a pattern of datachip
like :text/text and the url matches this pattern so it adds into the
list of data chip and when it tries to do mapping of the datachip it
cant find the appropriate data along with it which throws the error. I
changed the regular expression to avoid matching the URL.

## Testing Plan

Tested manually. If CI passes we are good.

## Output / Screenshots


https://github.com/allenai/olmo-api/assets/84343285/714d12fc-10a8-45a0-8314-d08819933908



### A11y
N/A