Set default AttributionRequest params, remove trailing slash from attribution endpoint (#133)
I was running into some issues with delimiters that may have been
because we weren't setting the parameters correctly. This PR sets those
parameters to make sure we don't have those issues.

The trailing slash was causing flask to want to redirect. Unfortunately,
the redirect redirects to http://olmo-api.allen.ai for some reason. This
is insecure and fails because of that.