[OEUI-123] use wildguard in the create_message_service (#139)
Addresses OEUI-123

This PR adds prompt checking for user requests. It builds off of the
work Yensung did in a previous PR, only changing some small things and
adding a call to `create_message_service`.

If a prompt is inappropriate, the API will respond with a 400 response
with the description `inappropriate_prompt`.