Make sure default model will be the first one in the list of that modal (#144)
Ref: https://allenai.atlassian.net/browse/OEUI-190
While adding the new Peteish model into the local `config.json `to make
sure it works i noticed that despite in the config i set the
`default_model` to be it but when the list renders on UI it still at the
last position. I made a tweak into to make sure for each `host` we will
want the `default_element` to be at the top instead of whatever the
position they are currently on for each `host`. I verified the model
work by submitting a question.
## Screenshot/Video 

![image](https://github.com/user-attachments/assets/fe524d05-5081-4349-9b94-d46ac913a938)

![image](https://github.com/user-attachments/assets/5e14527c-976e-4446-8d4a-8b81288ffe06)


https://github.com/user-attachments/assets/d62fb7c9-309c-4571-a8e4-fffd103692e2