Add custom field  to help marketing team to filter new user of Playground App (#154)
Closes: https://allenai.atlassian.net/browse/OEUI-251
Ashley has reached out and ask if there is a way to filter out user who
belong to Playground. This is not possible on their hubspot UI at all.
So one suggestion Hubspot recommended was creating a property when u
send out the contact payload. I spoke with Ashley and we agreed to use
the name create_by_playground_app as a way to indicate user first time
sign in & accept our T&C. When this field is created in hubspot contact
Ashley can use it to filter out all the new user of playground. The
value i just given it a text line since thats the easiest option they
were given us.
### Screenshot

![image](https://github.com/user-attachments/assets/ca4587ff-d307-4925-8213-55afef37afa3)

![image](https://github.com/user-attachments/assets/f38a31e6-cc21-44be-9ff7-bc10cb2222c9)