[OEUI-313] Save model info when creating a message (#155)
Closes OEUI-313

This PR makes it so that we save the model host and model ID along with
messages. This will allow us to show model names in the UI and ensure
that we use the correct index for CorpusLInk results.

The DB stuff in this project kills me. I'd love to move to an ORM and
migration generator at some point.

I tinkered with the message e2e tests to get them working with a0. Some
cases aren't covered anymore and we'll need to rework those in the
future!