[OEUI-357] Don't save messages or completions for anonymous users, don't allow them to share messages (#167)
Closes OEUI-357

This PR makes it so that messages for anonymous users have an expiration
time 1 day in the future. It also forces the messages to be private. We
also don't save completions for these messages at all to avoid potential
complications.

Message deletion will happen elsewhere. It may be an SQL script or it
may be attached to a cron job!

Tested through e2e test, curl, and @huytr1995 's PR
https://github.com/allenai/olmo-ui/pull/690