use gunicorn in our dockerfiles (#170)
i noticed that we weren't using a prod wsgi server when I was poking at
our logs yesterday. This fixes that! I copied dev.sh and start.sh from
the main skiff template.

I've tested locally with dev.sh and prod.sh and both are working well.

@codeviking I looked at the skiff template to see if there's anything
else we needed to do for this but didn't see anything. Do you remember
if there's anything else we need to do for this?