[OEUI-365] Add scheduled job to delete expired messages (#174)
Closes OEUI-365

This PR adds a dockerfile and k8s CronJob to delete expired messages
daily.

It uses a new Dockerfile, sql script, and sh script to do this. These
are built and deployed to our k8s cluster and scheduled to run daily.

I've tested locally but haven't tested in k8s, or tested the deploy.
Jsonnet builds it correctly.