[OEUI-358] [OEUI-373] Migrate T&C acceptance and messages to a logged in user (#181)
Depended on by https://github.com/allenai/olmo-ui/pull/715

This PR handles all the logic for migrating users from an anonymous
account to an authenticated one. it creates the `/v3/migrate-user`
endpoint. Not particularly RESTful, but it is what is is 😆

When someone hits this endpoint, we check if they're authenticated. If
they are, we move T&C acceptance and messages from the anonymous user to
the authed user.

<details>
<summary>
Response if user is new:
</summary>

```json
{
  "updated_user": null,
  "messages_updated_count": 0
}
```

</details>

<details>
<summary>
Response if new user exists:
</summary>

```json
{
  "updated_user": {
    "id": "user_W0U3F2E3C6",
    "client": "google-oauth2|106113864953374894702",
    "terms_accepted_date": "2024-11-18T23:12:11.782000Z",
    "acceptance_revoked_date": null
  },
  "messages_updated_count": 0
}
```

</details>

<details>
<summary>
Error if `anonymous_user_id` is missing:
</summary>

```json
{
  "error": {
    "code": 400,
    "message": "[{\"type\":\"string_type\",\"loc\":[\"anonymous_user_id\"],\"msg\":\"Input should be a valid string\",\"input\":null,\"url\":\"https://errors.pydantic.dev/2.7/v/string_type\"}]"
  }
}
```

</details>

<details>
<summary>
Error if not authed:
</summary>

```json
{
  "error": {
    "code": 401,
    "message": "Internal Server Error"
  }
}
```

</details>