[OEUI-491] Set up requests to MOLMo on Modal (#184)
Modal currently has a funky API so there was some custom code added for
it. We'll need to find out if we plan on standardizing it in the future
so we can adjust this.

I've set up some basic stuff to handle images for molmo specifically.
Again, we'll need to find out if we plan to standardize image inputs so
we can adjust for that in the future. The stuff in
create_message_service will need to change once we allow uploads in the
request!

Looking for feedback on the stuff in `ModalEngine` and `InferenceEngine`
since those are the changes I think should stick around. I was careful
not to generalize too much here since we'll be able to change those
whenever we want without big changes.