[OEUI-488] Create a v4 Message API (#185)
This PR sets up a `v4/message/stream` endpoint that will allow for image
uploads. I used `Pydantic` to create the request model to allow for easy
validation.

I got most of the validations added from the validation method in
`create_message_service`. All we're missing are things that need the DB!

I added
[`flask-pydantic-api`](https://github.com/adamsussman/flask-pydantic-api)
but I'm not sure it's worth it since we need to hack around it for
lists? It allows us to make an OpenAPI page for annotated endpoints but
the hackiness of lists really bugs me.

example request/response from what we have currently, response is just
the serialized request!:
request:
```
curl --request POST \
  --url http://localhost:8000/v4/message/stream \
  --header 'content-type: multipart/form-data' \
  --header 'x-anonymous-user-id: f941ecea-a872-4cf5-b076-4cf62ed8ebbc' \
  --form 'files=@/Users/taylorb/Desktop/Screenshot 2024-04-01 at 13.23.09.png,/Users/taylorb/Desktop/Screenshot 2023-08-22 at 2.33.07 PM.png' \
  --form 'content=what is this' \
  --form private=false \
  --form model_id=mm-olmo-uber-model-v4-synthetic \
  --form host=modal \
  --form max_tokens=2048 \
  --form n=1 \
  --form temperature=0.7 \
  --form top_p=1 \
  --form role=user \
  --form 'stop=\n' \
  --form stop=.
```
response:
```json
{
  "content": "what is this",
  "files": [
    "Screenshot 2024-04-01 at 13.23.09.png",
    "Screenshot 2023-08-22 at 2.33.07 PM.png"
  ],
  "host": "modal",
  "logprobs": null,
  "max_tokens": 2048,
  "model_id": "mm-olmo-uber-model-v4-synthetic",
  "n": 1,
  "original": null,
  "parent": null,
  "private": false,
  "role": "user",
  "root": null,
  "stop": [
    "\\n",
    "."
  ],
  "temperature": 0.7,
  "template": null,
  "top_p": 1
}
```

screenshot from Bruno: 

![image](https://github.com/user-attachments/assets/c1f24174-b187-4303-af76-ea796ee0c4a5)