match prod replicas, add a startup and liveness check (#215)
this matches the prod config for replicas. 

this also adds a liveness and startup check. the liveness check will run
after the startup check. if it fails 5 times in a row, it'll kill the
pod. this should allow us to automatically detect if any pod has failed
like it did last night and restart it automatically.