handle expired messages in our queries instead of relying on the job (#222)
I've disabled the expired message deletion job for now because I think
it may be causing some of our issues with 5xxs on the Playground. To
keep similar functionality, I'm adding checks for `expiration_time` in
our queries to get messages to show on the UI.

Tested locally through the API and UI.

I intentionally didn't add a check to the `get_by_root` query because it
deals with deleting things. Wasn't sure how we wanted to handle that one
so I left it alone!