Filter out documents that contain potential PII (#228)
Closes https://github.com/allenai/playground-issues-repo/issues/187

This PR makes it so that documents that contain what our regexes mark as
PII aren't shown to the user. This is the first step in our span mapping
pipeline so everything from here will work as normal.

It also sets up testing in Github Actions, except for our
`flatten_spans` test. I think we'll need to change how we get the config
or figure out how to mock it to be able to test with those in CI since
it ends up importing `cfg`, which wants the secret file to be there.

I'm not a huge fan of including another loop through the spans to handle
this but I also don't want to drive too deep into `flatten_spans` to
refactor this in there. We'll want to take another look at making this
performant in the future if it becomes a problem.

---------

Co-authored-by: Jiacheng Liu <liujch1998@gmail.com>