Stop initializing cfg when message or config are imported (#244)
We were having trouble in unit tests becuase `cfg` was getting imported
and initialized when it shouldn't have. I've fixed this by moving things
out of __init__.py files, which apparently run code whenever something
inside them is accessed. I did this for message as well since it was
also accessing `cfg`.

We'll want to see if we can put it back as something initialized at
runtime instead of build time but that's a problem for another day.

This fixes the unit test issues in
https://github.com/allenai/olmo-api/pull/243