don't leak our model configs in validation errors (#248)
Since we were including the input with our validation errors we were
leaking our model configs. This stops that, though it does make it so we
don't return the input along with the errors. We may be able to refactor
our file validation in the future to handle this!

example output:
```
{
  "error": {
    "code": 400,
    "message": "1 validation error for CreateMessageRequestFilesValidator\n  Value error, This model requires a file to be sent with the first message [type=value_error, input_value={'files': None, 'our_mode...e), 'has_parent': False}, input_type=dict]\n    For further information visit https://errors.pydantic.dev/2.10/v/value_error",
    "validation_errors": [
      {
        "loc": [],
        "msg": "Value error, This model requires a file to be sent with the first message",
        "type": "value_error",
        "url": "https://errors.pydantic.dev/2.10/v/value_error"
      }
    ]
  }
}
```