order spans by their start index, return start index to the UI (#259)
Addresses https://github.com/allenai/playground-issues-repo/issues/156

This PR ensures that spans are sorted by their `start_index` (left)
before being returned to the UI. This allows the UI to iterate through
the spans one at a time, only adding highlights for the first span.

It also returns the `start_index` to the UI. That's not essential, but
may be nice in the future or while debugging!