increase timeout to receive first token (#263)
Based on discussion from Slack:
https://allenai.slack.com/archives/C07530Y79Q9/p1743707972179919?thread_ts=1743706445.356419&cid=C07530Y79Q9

The timeout we set for the Tulu3 405B release was a little too short for
all models. This was causing Molmo to get "overloaded" errors when it
shouldn't have. @codeviking said that Molmo's TTFT p99 was 10s, so we're
going a little above that to play it safe.