Properly derive `remote_addr` from `Forwarded-For` (#272)
This fixes the issue we're currently debugging where the `remote_addr`
isn't using the `Forwarded-For` value.

Before this change we weren't returning the application that was wrapped
with `ProxyFix(app)`. This fixes that by overwriting `app`.

I considered modifying the code to `return ProxyFix(app, ...)`, but
that's a bigger shift that might have more significant side-effects.