improve our assessments by passing more info, reject request if score is 0.0 (#274)
addresses https://github.com/allenai/playground-issues-repo/issues/316

UI change: https://github.com/allenai/olmo-ui/pull/959

This PR improves our assessments by submitting more info (user IP and
agent) along with the token. It also adds a check to ensure the score is
above 0.0 and that the action is the same action that we expect. The
checks to verify a token are only performed if a user isn't logged in.