add a separate config to require the captcha_token to be present (#277)
This PR makes it so we can require the captcha token separately. If the
token is present we check it with google, otherwise we continue as
normal. This will allow us to filter tokens that score a 0 without
impacting users that don't load the recaptcha script right now.

We'll be looking into the recaptcha script loading more so we can
require the captcha_token for all users.