Ban IPs that were causing an attack (#279)
We had a bot attack the Playground this morning. As part of the changes
we added a lot of IP ranges to our ingress deny list. This adds those
IPs permanently so they aren't overwritten on a deploy.

This also removes the liveness check we added when we were having
trouble in another launch. This may have been causing excessive restarts
as the pods became overloaded. We'll continue to monitor stability of
the application with this disabled.