Make Create and Update model config endpoints accept either kind of model config (#292)
Our previous setup didn't allow these endpoints to actually accept
multi-modal models. Now they do!

The update fn is a little cursed but the other ways to update weren't
working for me so we're going with this for now.

Need to add tests to make sure we can create/update polymorphic models.