Get user permissions from whoami (#297)
Addresses https://github.com/allenai/playground-issues-repo/issues/338

This PR provides a way for the client to get the permissions a user has.
We can use it to show/hide UI elements according to their permissions.

Return value as an anonymous user:
```json
{
  "client": "f121b39f-f8b4-4508-bb2a-9e62b9097db3",
  "hasAcceptedTermsAndConditions": false,
  "id": null,
  "permissions": []
}
```

Return value as an admin:
```json
{
  "client": "google-oauth2|106113864953374894702",
  "hasAcceptedTermsAndConditions": true,
  "id": "user_M5D4H5X4M1",
  "permissions": [
    "read:internal-models",
    "write:model-config"
  ]
}
```

OpenAPI schema: 

![image](https://github.com/user-attachments/assets/30a179e2-e533-40c9-87e4-6ce7c5c76d4e)