Add Beaker Queue support (#309)
This gets the Beaker Queue model engine working (specifically for use
with vLLM-backed models). My local beaker-py wheel is included in the
build until my change to support queues in the Python client
(https://github.com/allenai/beaker/pull/6619) is released. Below is a
video using a "Beaker Test Queue" I added to the static config. The
output is a bit choppy, but I don't think this is related to the
queue/model. This queue is currently in Beaker dev, but I plan to
created a prod one after deploying Beaker tomorrow.


https://github.com/user-attachments/assets/1b3ce4d6-415f-447f-9e47-557d0de6eb38