Make new Thread endpoints that do message mapping for the client (#312)
Closes https://github.com/allenai/playground-issues-repo/issues/406

This PR is meant to make the UI's job easier. Since we're doing some
refactoring around message streaming anyway this seemed like a good
opportunity to move this functionality. It's basically a straight port
of what we had in v3/v4 messages but with `/threads` instead of
`/messages` or `/message`

There's some deeper changes we could make so that our logic doesn't
involve nesting these, but mapping at the controller level works for
now!