Split streaming handling to another file/function (#318)
Addresses https://github.com/allenai/playground-issues-repo/issues/423

This PR splits out the logic for streaming from a model, handling
exceptions, and getting metrics for the stream.

This is step 1 towards making a new endpoint we can use for
transcription. We'd like to deprecate that endpoint in the future when
we can get around using the DB for everything but I wanted to keep the
changes fairly simple for now. I played with making the DB optional for
the whole service but it was becoming a very large change.

@codeviking @mewil If you review this, could you make sure all the
metrics logic still makes sense? I'm having some trouble with
`StopIteration` right now but should have that fixed soon!