Add files only model config (#320)
Tested locally, need to run the migration in prod. May need to update
the UI before we start pushing files-only models to prod.

This PR adds a `files-only` model config type that we can use for
OLMoASR. This new model config type will let us do different validations
on the request and further define different properties for it in the
future.