Derive model name from id (#330)
When renaming models in olmo-api, we stopped being able to get responses
from Cirrascale's API. This is because we use the name as the model name
on the OpenAI API, and the models are hosted under static names that
Cirrascale set.

For now we work around this by deriving the model name from the ID that
is set by the caller.