Allow old auth0 issuer when validating tokens (#332)
Addresses https://github.com/allenai/playground-issues-repo/issues/405

This PR makes it so that our token validator can accept signed tokens
with an issuer of our current or old domain. This will prevent some 401s
while we transition.

After a couple days users that have logged back in will have the new
issuer and we'll be able to remove this code.

Tested locally by setting the auth0 domain on the client to the new and
old domains.