Fix message e2e tests (#342)
Updated the tests in the test_message.py file to get them passing.

FYI: This file takes about 3 minutes to run.

The tests did not account for a system message being added to the
conversation, so I updated them to accept this.

I Also fixed the config for the models, but this is still hardcoded so
could break again in the future.