Pydantic ai setup (#338)
Ticket: https://github.com/allenai/playground-issues-repo/issues/497

This PR add [Pydantic AI](https://ai.pydantic.dev/) as a inference
engine to call models. To toggle this on, set the feature flag
`enable_pydantic_inference` to true in the config.json file.

This PR also does a light refactor of the create_message_service, mostly
just breaking code into smaller files.


This new flow has not been tested with images. I need to find an open ai
spec model that can do so.


MessageChunk.logprobs are not support with the new pydantic code.
StreamMetrics are not support with the new pydantic code. 
Both are not currently reporting anything in the current backend with
cirrascale so, I would think this isn't an issue.


These changes pass our current e2e tests with the feature flag both on
and off.


NOTE: we can not turn the feature flag on in production until we move
all models over to cirrascale or ModalOpenAI

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>