Add cirrascale public endpoint (#345)
https://github.com/allenai/playground-issues-repo/issues/476

This PR adds cirrascale as an option both the existing inference engine
and the new pydantic engine.

### The image bug
I ran into a bug garbage response from the LLM when using the new
cirrascale endpoint, but i'm having issues reproduction this today. Here
are my findings and ideas on the issue.

* ~~The biggest change I made when implementing this was to go from
sending an image blob as part of the message content to sending a
file_url.~~ Reverted this seems to have fixed the issue.

* This could cause the issue if the file_url is not updated at the time
the llm is calling that URL. I check the image upload code and it
appears to wait for an upload, but hard to be 100% confident. These
seems like the most likely culprit.

* I'm pretty confident that the backend is sending a well formatted
response, because most of the time the endpoint works just fine.


#### Update 
After the change to not use url on first message, issues seems to have
resolved.


### Deploy note
Deploy will not succeeded until we add data["cirrascale"] to the config.


### API issue:
```
 Token usage limit exceeded for model 'Molmo-7B-D-0924'. Limit: 100000 tokens/minute.
 ```
 This isn't handled in the UI at the moment.

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>