Make e2e tests for threads, fix some issues (#355)
I was about to do some refactoring for how we save messages. I decided
it would be a good idea to make tests for threads endpoints before I do
that.

These are based off of test_message with some changes to hopefully make
them easier to read.

I also added the fake Pydantic AI model to our default config and use it
in e2e tests. This should make the tests faster and more consistent.