Tool calling setup (#353)
This PR adds tool calling to the backend using the pydantic-ai to call
said tools. It also further refactors create message.

After a tool is called, we return the tool response to the LLM (and the
ui) and generate a new assistant message in response. This means a
single call to the /thread endpoint can result in many assistant
messages being created.

I set a max of 10 assistant messages to avoid tool call run aways.



Closes
https://github.com/issues/assigned?q=is%3Aissue+state%3Aopen+archived%3Afalse+assignee%3A%40me+sort%3Aupdated-desc+org%3Aallenai&issue=allenai%7Cplayground-issues-repo%7C512,
https://github.com/allenai/playground-issues-repo/issues/510

- [x] Clean up streaming stats
- [x] Clean up TODOs
- [x] Discuss how we are going to store tools
- [x] Add first message token / Last message token?
- [x] Test without feature flag.

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>