Beaker use OpenAIStreamedResponse (#362)
## Description

Replace the creation of the streaming parts with passing directly to
`OpenAIStreamedResponse`. This _should_ handle all the parts for us.

I also stole a bunch of logic from the OpenAIModel for the message
mapping, as they also map the pydantic messages, and formatted them into
what I think we are looking for (with the exception of binary data; I
wasn't sure if turning that into a `data:base64` string was what we want
to send)

## Reviewer Notes

This is an updated version of #361 -- if this looks good more like the
right way to do this I will close that one, but I've left it for now.

This is working locally, but I don't have any way to test the extended
functionality (thinking, tool calls, etc) from beaker, as I don't think
the models I have those (yet?).