Make message repository (#367)
This PR completes the setup of Messages on SQL Alchemy. It also
refactors the create message flow to use the new SQL Alchemy types.

This PR moves tools from a json object on the message to a new table. 


- [x] remove feature flag. we can't really support the feature flag
without backwards support for the new tool call table. Probably can
remove.
- [x] fix type errors on errors placeholder
- [x] review todo items in code.

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>